feat: distillation path fixes, local WSL train notebook, Colab/eval updates
- Add Local_Train_Distill.ipynb (PROJECT_ROOT-relative paths for Linux/WSL)
- Colab train/eval notebooks and README updates
- Model, losses, train_distill, metrics, and tooling changes
- Exclude large LMDB/jpegio/build artifacts from commit

+++ models/fph.py
@@ -124,6 +124,6 @@ class FPH(nn.Module):
         self.conv0 = nn.Sequential(nn.Conv2d(in_channels=35, out_channels=256, kernel_size=8, stride=8, padding=0, bias=False),nn.BatchNorm2d(256, momentum=0.01),nn.ReLU(inplace=True),MBConvBlock(BlockArgs(num_repeat=repeats[0], kernel_size=3, stride=[1], expand_ratio=6, input_filters=in_channles[0], output_filters=in_channles[1], se_ratio=0.25, id_skip=True), global_params),MBConvBlock(BlockArgs(num_repeat=repeats[0], kernel_size=3, stride=[1], expand_ratio=6, input_filters=in_channles[1], output_filters=in_channles[1], se_ratio=0.25, id_skip=True), global_params),MBConvBlock(BlockArgs(num_repeat=repeats[0], kernel_size=3, stride=[1], expand_ratio=6, input_filters=in_channles[1], output_filters=in_channles[1], se_ratio=0.25, id_skip=True), global_params),)
 
     def forward(self, x, qtable):
-        x = self.conv2(self.conv1(self.obembed(x).permute(0,3,1,2).contiguous()))
+        x = self.conv2(self.conv1(self.obembed(x.long()).permute(0,3,1,2).contiguous()))
         B, C, H, W = x.shape
         return self.conv0(self.addcoords(torch.cat(((x.reshape(B,C,H//8,8,W//8,8).permute(0,1,3,5,2,4)*self.qtembed(qtable.unsqueeze(-1).unsqueeze(-1).long()).transpose(1,6).squeeze(6).contiguous()).permute(0,1,4,2,5,3).reshape(B,C,H,W),x), dim=1)))
